Fix getImageData error on _pixelSample
This happened when the recorder was hidden and _pixelSample was called
before the stream was loaded. Since videoWidth, videoHeight, clientWidth and
clientHeight  were equal to 0 the values being passed to _pixelSample were NaN.

If videoWidth and videoHeight are equal to 0 the pixelSample will always
return (0, 0, 0), even when we have clientWidth and clientHeight.  This
is because the video isn't loaded yet. So it doesn't make sense to get
pixel samples of something that isn't yet loaded.